bcakend/authservice/requirements.txt
Django
Pillow
djangorestframework
djangorestframework-jwt
redis
django-cors-headers